Add filter option

+++ backend/src/main/resources/db/migration/V1__init.sql
@@ -1,10 +1,12 @@
 -- Initial schema — compatible with SQLite and PostgreSQL.
 --
--- Uses BIGINT PRIMARY KEY to match Hibernate's Long id type (GenerationType.IDENTITY).
--- SQLite stores BIGINT as INTEGER affinity (same storage); Postgres uses its native BIGINT.
+-- Uses INTEGER PRIMARY KEY for ID columns. In SQLite, INTEGER PRIMARY KEY is an alias for the
+-- 64-bit rowid, which enables auto-increment and is what Hibernate's GenerationType.IDENTITY
+-- relies on via last_insert_rowid(). In PostgreSQL, INTEGER is a 32-bit int — sufficient for
+-- a family-scale app. A future migration can widen to BIGINT on Postgres if ever needed.
 
 CREATE TABLE IF NOT EXISTS users (
-  id              BIGINT       PRIMARY KEY,
+  id              INTEGER PRIMARY KEY,
   name            VARCHAR(255) NOT NULL UNIQUE,
   password_hash   VARCHAR(255) NOT NULL,
   is_admin        BOOLEAN      NOT NULL DEFAULT FALSE,
@@ -13,8 +15,8 @@ CREATE TABLE IF NOT EXISTS users (
 );
 
 CREATE TABLE IF NOT EXISTS gifts (
-  id               BIGINT       PRIMARY KEY,
-  owner_id         BIGINT       NOT NULL REFERENCES users(id),
+  id               INTEGER PRIMARY KEY,
+  owner_id         INTEGER      NOT NULL REFERENCES users(id),
   title            VARCHAR(255) NOT NULL,
   link             TEXT,
   price            VARCHAR(255),
@@ -29,10 +31,10 @@ CREATE TABLE IF NOT EXISTS gifts (
 );
 
 CREATE TABLE IF NOT EXISTS claims (
-  id               BIGINT    PRIMARY KEY,
-  gift_id          BIGINT    NOT NULL REFERENCES gifts(id)  ON DELETE CASCADE,
-  claimer_user_id  BIGINT    NOT NULL REFERENCES users(id)  ON DELETE CASCADE,
-  gift_date        DATE      NOT NULL,
+  id               INTEGER PRIMARY KEY,
+  gift_id          INTEGER  NOT NULL REFERENCES gifts(id)  ON DELETE CASCADE,
+  claimer_user_id  INTEGER  NOT NULL REFERENCES users(id)  ON DELETE CASCADE,
+  gift_date        DATE     NOT NULL,
   created_at       TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
   updated_at       TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
